Move Subable to the Substitutions module
